Minor refactor of resources
