+++ python/onvo/resources/accounts.py
@@ -1,7 +1,7 @@
-from onvo.resource import Resource
+from onvo.base_resource import BaseResource
 
 
-class Accounts(Resource):
+class Accounts(BaseResource):
     def list(self):
         return self.base_get("/accounts")
