+++ python/onvo/resources/embed_users.py
@@ -1,7 +1,7 @@
-from onvo.resource import Resource
+from onvo.base_resource import BaseResource
 
 
-class EmbedUsers(Resource):
+class EmbedUsers(BaseResource):
     def list(self):
         return self.base_get("/embed-users")
