+++ python/onvo/resources/teams.py
@@ -1,7 +1,7 @@
-from onvo.resource import Resource
+from onvo.base_resource import BaseResource
 
 
-class Teams(Resource):
+class Teams(BaseResource):
     def list(self):
         return self.base_get("/teams")
